rename metadata and types modules for clarity.
